Migrate nodes (#65)
* remove old percentile build

* update percentile entrypoint path

* update percentile entrypoint path

* save handler back in map

* migrate percentile to use refactor

* delete debug log

* delete old build path for group joiner

* edit entrypoint for group joiner

* refactor group joiner

* delete old q4 build

* refactor more than n reviews

* edit q4 entrypoint

* delete old more than n reviews

* delete old q3 build

* edit q3 entrypoint

* refactor top n reviews

* delete old version

* fix entrypoint for q1

* delete unused q1 build

* refactor games per platform
