+++ compose.yaml
@@ -51,7 +51,7 @@ services:
       - gateway
     deploy:
       mode: replicated
-      replicas: 3
+      replicas: 1
   decade-filter:
     image: tp1:latest
     entrypoint: /build/filter-decade
@@ -64,7 +64,7 @@ services:
       - gateway
     deploy:
       mode: replicated
-      replicas: 3
+      replicas: 1
   review-filter:
     image: tp1:latest
     entrypoint: /build/filter-score
@@ -76,7 +76,7 @@ services:
       - gateway
     deploy:
       mode: replicated
-      replicas: 3
+      replicas: 1
   language-filter:
     image: tp1:latest
     entrypoint: /build/filter-language
@@ -88,7 +88,7 @@ services:
       - gateway
     deploy:
       mode: replicated
-      replicas: 3
+      replicas: 8
   q1-partitioner:
     container_name: q1-partitioner
     image: tp1:latest
@@ -105,7 +105,7 @@ services:
   q1-count-1:
     container_name: q1-count-1
     image: tp1:latest
-    entrypoint: /games-per-platform
+    entrypoint: /build/games-per-platform
     environment:
       - RABBIT_IP=rabbitmq
       - PARTITION_ID=1
@@ -116,7 +116,7 @@ services:
   q1-count-2:
     container_name: q1-count-2
     image: tp1:latest
-    entrypoint: /games-per-platform
+    entrypoint: /build/games-per-platform
     environment:
       - RABBIT_IP=rabbitmq
       - PARTITION_ID=2
@@ -127,7 +127,7 @@ services:
   q1-count-3:
     container_name: q1-count-3
     image: tp1:latest
-    entrypoint: /games-per-platform
+    entrypoint: /build/games-per-platform
     environment:
       - RABBIT_IP=rabbitmq
       - PARTITION_ID=3
@@ -153,7 +153,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=games-Q2
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=game
     networks:
       - net
@@ -172,39 +172,13 @@ services:
       - net
     depends_on:
       - gateway
-  q2-top-2:
-    container_name: q2-top-2
-    image: tp1:latest
-    entrypoint: /top-n-historic-avg
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=2
-      - INPUT=games-Q2
-      - TOP_N=10
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q2-top-3:
-    container_name: q2-top-3
-    image: tp1:latest
-    entrypoint: /top-n-historic-avg
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=3
-      - INPUT=games-Q2
-      - TOP_N=10
-    networks:
-      - net
-    depends_on:
-      - gateway
   q2-joiner:
     container_name: q2-joiner
     image: tp1:latest
     entrypoint: /top-n-historic-avg-joiner
     environment:
       - RABBIT_IP=rabbitmq
-      - PARTITIONS=3
+      - PARTITIONS=1
       - INPUT=partial-Q2-joiner
       - TOP_N=10
     networks:
@@ -218,7 +192,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=games-Q3
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=game
     networks:
       - net
@@ -231,7 +205,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=reviews-Q3
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=review
     networks:
       - net
@@ -254,62 +228,10 @@ services:
   q3-top-1:
     container_name: q3-top-1
     image: tp1:latest
-    entrypoint: /top-n-reviews
+    entrypoint: /build/top-n-reviews
     environment:
       - RABBIT_IP=rabbitmq
       - PARTITION_ID=1
-      - N=5000
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q3-group-2:
-    container_name: q3-group-2
-    image: tp1:latest
-    entrypoint: /build/group-by
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=2
-      - GAME_INPUT=games-Q3
-      - REVIEW_INPUT=reviews-Q3
-      - OUTPUT=grouped-Q3-top
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q3-top-2:
-    container_name: q3-top-2
-    image: tp1:latest
-    entrypoint: /top-n-reviews
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=2
-      - N=5000
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q3-group-3:
-    container_name: q3-group-3
-    image: tp1:latest
-    entrypoint: /build/group-by
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=3
-      - GAME_INPUT=games-Q3
-      - REVIEW_INPUT=reviews-Q3
-      - OUTPUT=grouped-Q3-top
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q3-top-3:
-    container_name: q3-top-3
-    image: tp1:latest
-    entrypoint: /top-n-reviews
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=3
       - N=5000
     networks:
       - net
@@ -322,7 +244,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - TOP_N=5
-      - PARTITIONS=3
+      - PARTITIONS=1
     networks:
       - net
     depends_on:
@@ -334,7 +256,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=games-Q4
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=game
     networks:
       - net
@@ -347,7 +269,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=reviews-Q4
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=review
     networks:
       - net
@@ -367,41 +289,13 @@ services:
       - net
     depends_on:
       - gateway
-  q4-group-2:
-    container_name: q4-group-2
-    image: tp1:latest
-    entrypoint: /build/group-by
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=2
-      - GAME_INPUT=games-Q4
-      - REVIEW_INPUT=reviews-Q4
-      - OUTPUT=grouped-Q4-joiner
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q4-group-3:
-    container_name: q4-group-3
-    image: tp1:latest
-    entrypoint: /build/group-by
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=3
-      - GAME_INPUT=games-Q4
-      - REVIEW_INPUT=reviews-Q4
-      - OUTPUT=grouped-Q4-joiner
-    networks:
-      - net
-    depends_on:
-      - gateway
   q4-joiner:
     container_name: q4-joiner
     image: tp1:latest
-    entrypoint: /group-joiner
+    entrypoint: /build/group-joiner
     environment:
       - RABBIT_IP=rabbitmq
-      - PARTITIONS=3
+      - PARTITIONS=1
       - INPUT=grouped-Q4-joiner
       - OUTPUT=grouped-Q4-filter
     networks:
@@ -411,7 +305,7 @@ services:
   q4-filter:
     container_name: q4-filter
     image: tp1:latest
-    entrypoint: /more-than-n-reviews
+    entrypoint: /build/more-than-n-reviews
     environment:
       - RABBIT_IP=rabbitmq
       - N=5000
@@ -426,7 +320,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=games-Q5
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=game
     networks:
       - net
@@ -439,7 +333,7 @@ services:
     environment:
       - RABBIT_IP=rabbitmq
       - INPUT=reviews-Q5
-      - PARTITIONS=3
+      - PARTITIONS=1
       - TYPE=review
     networks:
       - net
@@ -459,41 +353,13 @@ services:
       - net
     depends_on:
       - gateway
-  q5-group-2:
-    container_name: q5-group-2
-    image: tp1:latest
-    entrypoint: /build/group-by
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=2
-      - GAME_INPUT=games-Q5
-      - REVIEW_INPUT=reviews-Q5
-      - OUTPUT=grouped-Q5-joiner
-    networks:
-      - net
-    depends_on:
-      - gateway
-  q5-group-3:
-    container_name: q5-group-3
-    image: tp1:latest
-    entrypoint: /build/group-by
-    environment:
-      - RABBIT_IP=rabbitmq
-      - PARTITION_ID=3
-      - GAME_INPUT=games-Q5
-      - REVIEW_INPUT=reviews-Q5
-      - OUTPUT=grouped-Q5-joiner
-    networks:
-      - net
-    depends_on:
-      - gateway
   q5-joiner:
     container_name: q5-joiner
     image: tp1:latest
-    entrypoint: /group-joiner
+    entrypoint: /build/group-joiner
     environment:
       - RABBIT_IP=rabbitmq
-      - PARTITIONS=3
+      - PARTITIONS=1
       - INPUT=grouped-Q5-joiner
       - OUTPUT=grouped-Q5-percentil
     networks:
@@ -503,7 +369,7 @@ services:
   q5-percentile:
     container_name: q5-percentile
     image: tp1:latest
-    entrypoint: /90-percentile
+    entrypoint: /build/percentile
     environment:
       - RABBIT_IP=rabbitmq
       - PERCENTILE=90
